c language files renaming, still needed
