+++ LadderApp/PLC/addresses.h
@@ -9,7 +9,7 @@
 //====================================================================================
 // Includes
 //====================================================================================
-#include "definicao.h"
+#include "definitions.h"
 
 //====================================================================================
 // Variáveis Globais
