+++ LadderApp/PLC/functions.h
@@ -8,8 +8,8 @@
 //====================================================================================
 // Includes
 //====================================================================================
-#include "definicao.h"
-#include "setuphardware.h"
+#include "definitions.h"
+#include "hardwaresetup.h"
 
 
 //====================================================================================
