+++ LadderApp/PLC/hardwaresetup.c
@@ -7,7 +7,7 @@
 //====================================================================================
 // Includes
 //====================================================================================
-#include "setuphardware.h"
+#include "hardwaresetup.h"
 
 //====================================================================================
 // Rotina de configuração do sistema
