+++ LadderApp/PLC/hardwaresetup.h
@@ -10,8 +10,8 @@
 //====================================================================================
 // Includes
 //====================================================================================
-#include "definicao.h"
-#include "enderecos.h"
+#include "definitions.h"
+#include "addresses.h"
 
 //====================================================================================
 // Prototipos
