+++ LadderApp/PLC/ladderprogram.h
@@ -10,8 +10,8 @@
 //====================================================================================
 // Includes
 //====================================================================================
-#include "definicao.h"
-#include "setuphardware.h"
+#include "definitions.h"
+#include "hardwaresetup.h"
 #include "functions.h"
 
 //====================================================================================
